pythonmachinetest2/requirements.txt
Django>=4.0
djangorestframework